pyproject.toml
[tool.poetry]
name = "twitter-clone-backend"
version = "0.1.0"
description = "Twitter Clone API"
authors = ["pedrosimiao <pedrolumiere@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
djangorestframework = "^3.16.0"
django-debug-toolbar = "^5.0.1"
django-extensions = "^3.2.3"
djangorestframework-simplejwt = "^5.4.0"
pillow = "^11.1.0"
factory-boy = "^3.3.3"
django-cors-headers = "^4.7.0"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.9.0"
gunicorn = "^23.0.0"
gitpython = "^3.1.44"
django-environ = "^0.12.0"
django-filter = "^25.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
